webapp/requirements.txt
Django>=6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
plotly>=5.0
pandas>=2.0
pyarrow>=12.0
boto3>=1.28
python-dotenv>=1.0
duckdb>=0.9.0
gunicorn>=21.0
whitenoise>=6.0
authlib>=1.3.0
requests>=2.31.0
